requirements.txt
groq==0.4.2
python-dotenv==1.0.0
sentence-transformers==2.2.2
psycopg2-binary==2.9.9
numpy==1.24.3